Fix #30

+++ src/main/java/hk/ust/cse/comp4111/handler/BookDeleteRequestHandler.java
@@ -13,6 +13,8 @@ import java.util.Map;
 
 public class BookDeleteRequestHandler extends ServerRequestHandler {
 
+    private static final String PATH_PREFIX = "/BookManagementService/books/";
+
     public BookDeleteRequestHandler() {
     }
 
@@ -23,9 +25,12 @@ public class BookDeleteRequestHandler extends ServerRequestHandler {
             return;
         }
 
-        String[] temp = path.split("/");
-        temp = temp[temp.length - 1].split("\\?");
-        String idFromURL = temp[0];
+        if (!path.startsWith(PATH_PREFIX)) {
+            response.setStatusCode(HttpStatus.SC_INTERNAL_SERVER_ERROR);
+            return;
+        }
+
+        String idFromURL = path.substring(PATH_PREFIX.length());
 
         try {
             boolean success = BookService.getInstance().deleteBook(Integer.parseInt(idFromURL));
